requirements.txt
# Core ML Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Vector Store
faiss-cpu>=1.7.0

# LLM
groq>=0.4.0

# Web App
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.12.0
pydantic-settings>=2.12.0
requests>=2.32.0
httpx>=0.28.0
aiohttp>=3.13.0

# Database & Storage
SQLAlchemy>=2.0.0
PyYAML>=6.0.0

# File Handling
PyMuPDF>=1.26.0
pypdf>=6.5.0

# Development & Testing
pytest>=9.0.0
build>=1.3.0
setuptools>=80.0.0
wheel>=0.45.0

# Additional Tools
tqdm>=4.67.0
click>=8.3.0
rich>=14.2.0
colorama>=0.4.6
python-dateutil>=2.9.0
huggingface-hub>=0.36.0
joblib>=1.5.0